Refactor to "composables"
